Simplify the groups.
It turns out that the other changes have made it such that we can get
away with an interface for the group definitions, since all of the
classes involved now have AbstractNamedActivity as a common parent.
Convert GenericActivityGroup to an interface and remove it's subclasses,
moving the implementation (which is pretty much there anyways) to the
classes which implment the interface.