Refactor the enrolment handling classes.
All of the enrolment editing functionality is moving to the Enrolment
factory/manager.  As a result all of the mutator methods that were in
EnrolledUser can move to protected versions AbstractEnrolment, with the
protected vversioned that were in EnrolledCourse being removed.  With
the migration, EnrolledCourse nolonger has any purpose so delete it and
rename Abstract Enrolment to EnrolmentData, to reflact the change.
Rename EnrolledUser to UserEnrolmentData for consistantcy.