Re-factor manager instantiation.
Remove the DomainModelType Enum, and replace it will passing the class
object to the Domain model, as that works better with the generic
factory method.  Update the constructor and static factory (convenience)
method to reflect this change.  Update the documentation to reflect this
changes and add a means to pass in the query and builder, and setup the
logs.