Remove the Generics from the Activity and Log classes
They never did function as expected, and they will get in the way with
the element factories.  To keep them in the element factories would
require a cast which introduces a runtime check, so it better (in terms
of documentation) to do the runtime check explicitly.

Also completely document the classes and fix some minor issues.