Make the Activity and SubActivity Builders hierarchical
Adjust the builder implementation such that the AbstractActivityBuilder and
AbstractSubActivityBuilder can be subclassed to a builder that creates
an Element implementing any subclass of Activity or SubActivity
respectively.

Change all of the set methods to return void instead of the builder
instance, since implementing hierarchical builder that would always
return a builder instance of the right type is more trouble than it's
worth.

Fix some incomplete and incorrect documentation.

Add the missing datastore checks to the LogEntryBuilder.