Make getMetaData return definitions only.
Previously, getCreator and getMetadata did the same thing, both looked
up the default implementation class (with one argument) then went and
found the Metadata (or Creator) instance for that class.

getMetadata now does not look up the default implementation class, and
the two argument version has been removed.  It will now only return
Definitions (or Implementations, if that is what was requested).