Add a setid method to the idgenerators and a Pass-Through idgenerator
There are situations where an Element is initialized with an ID number
and that number needs to be retained.  The common situation is when an
Element is loaded from a database, has a transform applied to it, and
the result is cached.  In that situation we want the ID numbers for the
original and transformed elements to match so that we can easily find
the transformed version when dealing with other Elements that reference
the original.

The best/most OO way to do this is to have an IdGenerator that passes
through the original ID.  The Pass-Through generator requires that the
IdGenerators write the ID number into the Element.