Restructure the Query logic
It turns out that ARIN's database is not consistent.  To make the query
work and give reasonable results, take a recursive-first approach.  If a
network as a parent entry recursive process it, until there is not
parent entry.  For the result choose the highest level with a
consistent (and non-empty) organization name.