Use the datastore ID the Enrolment's equals and hashcode
Enrolment has a problem, it is supposed to be unique with respect to the
user and the course, but the user isn't available for comparison.  To
get around the problem, we are using the datastore ID as a stand in, as
it will make enrolment's suitably different.

To give us the expected equals behaviour, the equalsUnique method has
been added which performs the comparison with out using datastore
details.

The equalsAll method has been added as well to perform the comparison
using all of the fields (except the data-store details)