Catch any cases where the activity type isn't registered
If the ActivityType isn't registered then everything will be fine until
the NamedActivityBuilder tries to create the instance of the target class.
Then there will be a ClassCastException because the target class was
GenericActivity, as the real target didn't have metadata loaded.  Catch
the problem when the builder is created an issue a better error message.