Move the Id, course and type up to Activity.
Hibernate barfs with Property access and defensive copies.  Apparently
eclipselink handles defensive copies with property access properly.
However, it barfs if the methods for the primary key are abstract
(undocumented).  So move these three up as they need to be defined on
Activity for the course database.  If eclipselink doesn't work, then
we'll have to revert this change.