Split the common Activity data out into a separate class
This is a work-around for issues with the JPA implementation.  In short,
while JPA officially supports access by property, it doesn't actually
work with most implementations breaking in different ways.  To use field
access, Activity has to be split so that we can avoid issues with moodle
and it repeated id numbers.