Use the TranslationTable only when building LogReferences
This is a first stage fix for a database performance issue when creating
LogReferences.  The DataStoreProxy checks redundant instances when
inserting the LogReference.  Since LogReference is defined in terms of
the LogEntry, normally a database query would be used for the check.
For LogReference this query is extremely expensive (polynomial time or
worse).  It is also redundant, as the LogEntry can be checked to ensure
that there is no redundancy.  So force the use of the TranslationTable
only to avoid the problem.

This problem should be permanently fixed in a refactor of the
DataStoreProxy.