Use separate caches for id and type in ActivityConverter
The ActivityConverter cached Activities using the Id and the type,
however these are independent, with the type only coming into play if
the id is zero.  Furthermore a separate cache needed to be kept to map
the non-zero id's to the id/type key to deal with situation where moodle
was not consistent.

With the pending removal of apache.common.collections, multikey needs to
be replaced.  Going to a separate caches for id an type simplifyes
things using the same number of maps, and means that the value class
isn't required.