Remove the MetaData Mutator
Upon further consideration the Mutator is dangerous, as there is a
straight forward method for getting a Mutator outside of the very narrow
case where it was intended.  The Mutator is only needed for two use
cases:  Setting the DomainModel and Id number in the Persister, and
setting the DomainModel in the Query (JPA only).  Getting a Mutator into
the Query is difficult as the metadata isn't available on the Query
creation path.  It's easier and safer to create the method reference when it is
needed.