Merge Persister into the DomainModel
With the integration of the Builders into the Element and the common
build method, the persister is redundant.  Place most of it's
functionality into the DomainModel, replacing the the getDataStore
method to allow the DomainModel to completely encapsulate the DataStore.

Remove the getSelectors and getProperties methods from the Element, as
they aren't useful without the type parameters, which need to be
stripped due to inheritance.  Replace these methods with protected
connect and disconnect methods to allow the DataStore to have the
Elements connect and disconnect their relationships.

Also update the getQuery methods to reflect the changes to the
datastore.