Add Definitions to the Moodle Element classes
The Profile now requires the class to have a definition in order to be
added to the Profile.  So Add Definitions to the Moodle-specific Element
implementations.  Since these classes are intended to be loaded from the
Moodle database, and would require specialized builders, which would
never be used, the constructors which create the Element from the
builder throw UnsupportedOperationException.