Remove the recusive definition from the MetaData
If MetaData components from parent classes need to be referenced, then
the Element class will be responsible for performing the recursion.
This significantly simplifies the MetaData.

Use ArrayLists instead of HashSets to store the Metadata components.
ArrayList takes less memory, and is more appropriate since all access is
an iteration.