Element's must be from different DomainModel's for a table lookup
The translation table only applied when the queried element instance is
from a different domain model than the one which the retriever
references.  Since the Retriever is called just prior to performing the
insert, the query element and the domain model can be the same.  In this
case only perform the query (if appropriate) and skip the table.