Remove equals and hashcode
With the translation table strictly requiring that elements be
identical, MoodleActivityReference will no longer be associated with
other ActivtyReference implementations.  We don't really care about the
instanceid as it should be transient (a factor for just long enough to
load the actual activity), so remove the equals and hashCode methods.