Use the MetaData to store the dependencies
When sorting the Elements during synchronization the entire dependency
tree needs to be analysed for each element.  For Example LogEntry does
not depend on Role.  However LogEntry depends on Enrolment which depends
on Role.  So Role must sort ahead of LogEntry when the two are compared,
which requires a recursive analysis of the dependencies.

The only time that the dependency tree is known to an Element class is
when its MetaData is being constructed.  MetaData and Class
initialization is structured such that the dependencies for an Element
are loaded before the element.  So the MetaData can easily build and
store a complete list of the dependencies (direct and indirect), which
can be queried during the sort.