Remove the enrolmentQuery from User.Builder's parameters
User.enrol hardcodes the query to use to ensure that the specified
enrolment is not already associated with another user.  So it doesn't
make sense for it to be a parameter to the builder.

The query functionality from User.enrol was moved to a new private
method which is called by User.enrol and the Builder.