Revert the removal of the ID from Enrolment.equals
To find the associated user's Enrolments need to be unique, which
requires the use of the ID number.  Unfortunately the IDnumber is not
stable across DomainModel instances, so the same Enrolment will have
different ID's in different DomainModels.

The TranslationTable requires associated elements to be identical
(across DomainModels).  Enrolment gets around the problem by not
including the ID number in equalsAll.

The difference in the Enrolment ID's makes LogEntry's not identical
across datastores, which screws up the translation table.  To work
around the problem, implement equalsAll on LogEntry and define all
relationships in equalsAll using the equalsAll method for the associated
elements.