Create a post-Build hook for User
User needs to add any new enrolments, once the user instance has been
built or retrieved from the DataStore.  The post-insert hook is only
applied to new instances, the User builder needs to act on pre-existing
instances as well.  So add a post build hook that will act on
everything.