Use dependency level-ordering to determine the sort order
Comparing the dependency trees for two classes to determine the sort
order doesn't work in cases where there are two un-related classes (i.e.
Network, and Role).  To correct this we need to look at the entire
dependency graph.

The dependencies between Element classes can be modelled as a directed
acyclic graph.  A level-ordering of the graph can be used to to order
the Element classes such that dependencies are processed first.  For
classes in the same level, order is determines based upon the class
name.

The level is calculated when the MetaData instance for the class is
created, by adding one to the level of the highest level dependency.
Classes which have no dependencies are level zero.