Use a selective-evict batching algorithm for synchronization
The JPA EntityManager performs poorly if it has a lot of elements in
its persistence context.  Based on testing it seems that query
performance has a time complexity of O(n^2), or worse, relative to the size of
the persistence context.  The recommended solution to the issue is to
break large operations into small batches.  However, this dataset really
doesn't want to be broken into batches.

To process the contents of the domain model, the synchronizer uses a
selective eviction algorithm.  The algorithm tracks all of the
dependents for a given Element.  Once all of the dependents have been
processed the Element is scheduled for eviction from the persistence
context and the translation table.  The evictions are processed after
the transaction is committed at the end of the batch.

Due to the inter-related nature of the contents of the domain model, a
fixed batch size is ineffective.  Ideally the batch size would be
smaller at the end of the run and larger during the middle of the run.
To achieve this a dynamic batch size is used, based on the size of the
evict queue.  Two parameter control when the evict queue is processed: a
loading factor and a minimum queue length.  The loading factor states
that the transaction should be committed and the evictions processed
once the evict queue length reaches a specified percentage of the
estimated size of the persistence context.  To prevent the evictions
from dominating the processing at the end of the run when the
persistence context will be small, a minimum queue size is set.

Currently the loading factor is set to 20% and the minimum queue length
is 100 elements.  With a run of about 1.1M elements the estimated
persistence context size peaked at about 150k elements.  The performance
was generally acceptable, but the performance at the end of the run was
noticeably slower than at the beginning of the run.