Add and ID field to Grade
Add an Id field to grade to act as the primary key.  Ideally grade would
use a composite primary key based on the activity and enrolment id's.
However such an approach would require that all of the ID fields become
Object, rather then the current Long.  That change would have wide
effects and make type safety impossible.  So use a generated key for
grade as a compromise to JPA.